server/requirements.txt
google-genai>=1.0.0
gTTS>=2.5.0
librosa>=0.11.0
numpy>=1.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0
soundfile>=0.13.0
